Added comment regarding error handling for alt text
I initially had error handling for when altText was more than 120 characters, but I struggled a lot to get it to work. It prompted me with Error code 400 no matter how I tried to implement it. Therefore I decided to add information about the maximum limit in the form, and make it optional as I was running out of time. If altText is not added, altText will be set to blog post title